Factor out bidform
